pyproject.toml
[project]
name = "varick"
version = "0.1.0"
description = "AI agent for Accounts Payable automation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "pydantic-ai>=0.0.49",
    "python-dotenv>=1.0",
    "pytest>=9.0.2",
]

[project.scripts]
gl = "main:cli"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
